ffs: add initial ffs support to inspect max order on iomap iter write
The patch titled "filemap: Allow __filemap_get_folio to allocate large folios"
let's us prioritize higher order folios if possible on the buffered-IO
write path for iomap. It uses the length of the write as the max order
and we cap at MAX_PAGECACHE_ORDER.

Even after this though we also check to see if the index of the write
position is aligned to the target order, if not we override the order
used with __ffs(index). This let's us verify the expected order used
manually.

We import the kernel's __ffs.h.

We demo this below with order, 2, ie 16k.

Demo 1:

At order 2, offset of 16384 things are aligned:

./ffs --order 2 --offset 0 --count 3
           pos: 0    index: 0  ffs(0) : 63
           pos: 1    index: 0  ffs(0) : 63
           pos: 2    index: 0  ffs(0) : 63

Demo 2:

At order 2, offset of 0 things are aligned:

./ffs --order 2 --offset $((16384)) --count 3
           pos: 16384    index: 4  ffs(4) : 2
           pos: 16385    index: 4  ffs(4) : 2
           pos: 16386    index: 4  ffs(4) : 2

Demo 3:

At offset 16384-3 we're unaligned to start, so if any initial
pos used is in between 16381 up to 16383 the index would be unaligned
to the target order desired as the index is 3, so we end up using
on iomap order 0.

3 & (4-1) = 0b11  unaligned
4 & (4-1) = 0     aligned

./ffs --order 2 --offset $((16384-3)) --count 6
 unaligned pos: 16381    index: 3  ffs(3) : 0
 unaligned pos: 16382    index: 3  ffs(3) : 0
 unaligned pos: 16383    index: 3  ffs(3) : 0
           pos: 16384    index: 4  ffs(4) : 2
           pos: 16385    index: 4  ffs(4) : 2
           pos: 16386    index: 4  ffs(4) : 2

Demo 4:

At offset 16384 + 4096 - 3, we have unalignment at index 5
and so order 0 is used for the target order.

/ffs --order 2 --offset $((16384 + 4096 - 3)) --count 6
           pos: 20477    index: 4  ffs(4) : 2
           pos: 20478    index: 4  ffs(4) : 2
           pos: 20479    index: 4  ffs(4) : 2
 unaligned pos: 20480    index: 5  ffs(5) : 0
 unaligned pos: 20481    index: 5  ffs(5) : 0
 unaligned pos: 20482    index: 5  ffs(5) : 0

Demo 5:

At offset $((4096*2)) if we write 1000000 bytes we check the
different unique unaligned order used:

./ffs --order 2 --offset $((4096*2)) --count 1000000 | grep unaligned | awk '{print $8}'| sort | uniq
0
1

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>